fix: organise probe pages and components
