+++ src/components/ProbeTokenModal/index.ts
@@ -1,0 +1,1 @@
+export { ProbeTokenModal } from './ProbeTokenModal';
