+++ src/data/useProbes.ts
@@ -1,6 +1,10 @@
 import { useContext, useEffect, useState } from 'react';
+import { useAsyncCallback } from 'react-async-hook';
+import { AppEvents } from '@grafana/data';
+import appEvents from 'grafana/app/core/app_events';
 
 import { type Probe } from 'types';
+import { FaroEvent, reportError, reportEvent } from 'faro';
 import { InstanceContext } from 'contexts/InstanceContext';
 import { useTrigger } from 'hooks/useTrigger';
 
@@ -41,4 +45,138 @@ export function useProbes() {
     probes,
     refetchProbes,
   };
+}
+
+export type CreateProbeResult = {
+  probe: Probe;
+  token: string;
+};
+
+export function useCreateProbe() {
+  const { instance } = useContext(InstanceContext);
+  const event = FaroEvent.CREATE_PROBE;
+
+  const { execute, error, loading, result } = useAsyncCallback(
+    async (probe: Probe, onSuccess: (res: CreateProbeResult) => void) => {
+      if (!instance.api) {
+        throw new Error('Not connected to the Synthetic Montoring datasource');
+      }
+
+      const info = await instance.api.addProbe({
+        ...probe,
+        public: false,
+      });
+
+      reportEvent(event);
+      appEvents.emit(AppEvents.alertSuccess, [`Created probe ${info.name}`]);
+      onSuccess(info);
+    }
+  );
+
+  useOnError({ event, error, alert: `` });
+
+  return {
+    result,
+    onCreate: execute,
+    error,
+    loading,
+  };
+}
+
+export type UpdateProbeResult = {
+  probe: Probe;
+};
+
+export function useUpdateProbe() {
+  const { instance } = useContext(InstanceContext);
+  const event = FaroEvent.UPDATE_PROBE;
+
+  const { execute, error, loading, result } = useAsyncCallback(
+    async (probe: Probe, onSuccess: (res: UpdateProbeResult) => void) => {
+      if (!instance.api) {
+        throw new Error('Not connected to the Synthetic Montoring datasource');
+      }
+
+      const info = await instance.api.updateProbe(probe);
+
+      reportEvent(event);
+      appEvents.emit(AppEvents.alertSuccess, [`Updated probe ${info.probe.name}`]);
+      onSuccess(info);
+    }
+  );
+
+  useOnError({ event, error });
+
+  return {
+    result,
+    onUpdate: execute,
+    error,
+    loading,
+  };
+}
+
+export function useDeleteProbe() {
+  const { instance } = useContext(InstanceContext);
+  const event = FaroEvent.DELETE_PROBE;
+
+  const { execute, error, loading, result } = useAsyncCallback(async (probe: Probe, onSuccess: () => void) => {
+    if (!instance.api) {
+      throw new Error('Not connected to the Synthetic Montoring datasource');
+    }
+
+    await instance.api.deleteProbe(probe.id!);
+
+    reportEvent(event);
+    appEvents.emit(AppEvents.alertSuccess, [`Deleted probe ${probe.name}`]);
+    onSuccess();
+  });
+
+  useOnError({ event, error });
+
+  return {
+    result,
+    onDelete: execute,
+    error,
+    loading,
+  };
+}
+
+export function useResetProbeToken(probe: Probe, onSuccess: (token: string) => void) {
+  const { instance } = useContext(InstanceContext);
+  const event = FaroEvent.RESET_PROBE_TOKEN;
+
+  const { execute, error, loading, result } = useAsyncCallback(async () => {
+    if (!instance.api) {
+      throw new Error('Not connected to the Synthetic Montoring datasource');
+    }
+
+    const { token } = await instance.api.resetProbeToken(probe);
+
+    reportEvent(event);
+    onSuccess(token);
+  });
+
+  useOnError({ event, error, alert: `Failed to reset probe token` });
+
+  return {
+    result,
+    onResetToken: execute,
+    error,
+    loading,
+  };
+}
+
+type useOnErrorProps = {
+  event: FaroEvent;
+  error?: Error;
+  alert?: string;
+};
+
+function useOnError({ event, error, alert }: useOnErrorProps) {
+  useEffect(() => {
+    if (error) {
+      reportError(error.message, event);
+      alert && appEvents.emit(AppEvents.alertError, [alert]);
+    }
+  }, [error, event, alert]);
 }
