+++ src/page/EditProbe/index.ts
@@ -1,0 +1,1 @@
+export { EditProbe } from './EditProbe';
