+++ src/page/NewProbe/index.ts
@@ -1,0 +1,1 @@
+export { NewProbe } from './NewProbe';
